vtkAMReXGridReaderInternal: fix signed/unsigned comparison
(cherry picked from commit 1be2c0821d1c72809c602b41612344a66f44f458)

+++ IO/AMR/vtkAMReXGridReaderInternal.cxx
@@ -920,7 +920,7 @@ void vtkAMReXGridReaderInternal::GetBlockAttribute(
 
       // read every component of the variable into the buffers vector
       std::string attributeName(attribute);
-      size_t nComps = this->Header->parsedVariableNames[attributeName].size();
+      int nComps = static_cast<int>(this->Header->parsedVariableNames[attributeName].size());
       std::vector<std::vector<char>> buffers(nComps);
       for (int i = 0; i < nComps; ++i)
       {
